refactor: extract showToast
